+++ src/providers/ZIndexDiagnosticsProvider.ts
@@ -6,8 +6,12 @@ import { Uri } from "vscode";
 import {
   diagnosticsCollection,
   INEFFECTIVE_Z_INDEX_CODE,
+  PARTICIPATES_IN_PARENT_CODE,
 } from "../contants/globals";
-import { isIneffectiveZIndexDeclaration } from "../helpers/findStackingContexts";
+import {
+  isIneffectiveZIndexDeclaration,
+  isInParentStackingContext,
+} from "../helpers/findStackingContexts";
 import { nodeRange } from "../helpers/nodeRange";
 import { Logger } from "../helpers/logger";
 import debounce from "lodash.debounce";
@@ -22,6 +26,9 @@ export class ZIndexDiagnosticsProvider {
   public async provideZIndexDiagnostic(
     document: vscode.TextDocument,
   ): Promise<void> {
+    if (!["css", "scss"].includes(document.languageId)) {
+      return;
+    }
     try {
       const text = document.getText();
       const result = await postcss().process(text, {
@@ -31,12 +38,17 @@ export class ZIndexDiagnosticsProvider {
       const diagnostics: vscode.Diagnostic[] = [];
       result.root.walkDecls("z-index", (decl) => {
         const isIneffective = isIneffectiveZIndexDeclaration(decl);
+        const isInParentContext = isInParentStackingContext(decl);
+        let diagnosticMessage = `The \`z-index\` value of ${decl.value} is ineffective because the rule does not create a new stacking context.`;
 
         if (isIneffective) {
           const range = nodeRange(decl);
+          if (isInParentContext) {
+            diagnosticMessage += ` The rule participates in the parent stacking context.`;
+          }
           const diagnostic = new vscode.Diagnostic(
             range,
-            dedent`The \`z-index\` value of ${decl.value} is ineffective because the rule does not create a new stacking context.`,
+            dedent`${diagnosticMessage}`,
             vscode.DiagnosticSeverity.Warning,
           );
           diagnostic.code = INEFFECTIVE_Z_INDEX_CODE;
@@ -46,8 +58,14 @@ export class ZIndexDiagnosticsProvider {
 
       this.cache.set(document.uri.toString(), diagnostics);
       diagnosticsCollection.set(document.uri, diagnostics);
-    } catch (e) {
-      Logger.error(`Error in provideZIndexDiagnostic: ${e}`);
+    } catch (e: any) {
+      if (e.name === "CssSyntaxError") {
+        Logger.error(
+          `CssSyntaxError in provideZIndexDiagnostic: ${e.reason} at line ${e.line}, column ${e.column}`,
+        );
+      } else {
+        Logger.error(`Error in provideZIndexDiagnostic: ${e}`);
+      }
     }
   }